// Mobile/types/event.ts
export interface EventTag {
    tagName: string;
}

/** Match BE EventEventLocationDTO */
export interface EventLocation {
    id?: string;
    name?: string;
    address: string;
    province?: string;
    district?: string;
    ward?: string;
    zipcode?: string;
    latitude?: number;
    longitude?: number;
    contactEmail?: string;
    contactPhone?: string;
}

/** Match BE EventDTO */
export interface Event {
    id: string;
    name: string;
    slug: string;
    subtitle?: string;
    description: string;
    tags?: EventTag[];
    startTime: string;
    endTime: string;
    openTime?: string;
    closedTime?: string;
    status: EventStatus;
    thumbnailUrl?: string;
    bannerUrl?: string;
    ticketMapUrl?: string;
    ageRestriction: number;
    category?: Category;
    organizer?: Organizer;
    locations?: EventLocation[];
    ticketTypes?: TicketType[];
}
// ...
export interface Category {
    id: string;
    name: string;
    slug: string;
    description?: string;
    iconUrl?: string;
    status: CategoryStatus;
    parentCategory?: Category | null;
    subCategories?: Category[];
    isDeleted?: boolean;
}
// ...
export interface Organizer {
    id: string;
    name: string;
    slug: string;
    description?: string;
    logoUrl?: string;
    bannerUrl?: string;
    email: string;
    phone?: string;
    websiteUrl?: string;
    facebookUrl?: string;
    instagramUrl?: string;
    tiktokUrl?: string;
    address?: string;
    isVerified: boolean;
    status: OrganizerStatus;
    createdAt?: string;
    updatedAt?: string;
    events?: Partial<Event>[];
}